79_GPTNotepad/requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-generativeai
bs4
requests
python-docx
PyPDF2
